requirements.txt
numpy
pandas
scikit-learn
tensorflow
flask
flask-restful
shapely
osmnx
geopandas
folium
matplotlib
networkx
python-dotenv
rtree
requests
seaborn
jupyter